refactor bookitem
